# apps/api/src/routes/scans.ts
import type { FastifyInstance } from "fastify";
import { analyzeScanRequestSchema, confirmScanRequestSchema } from "@logmyplate/contracts";
import { decideScanQuota } from "@logmyplate/domain";
import type { AppRepository } from "../repositories/app-repository.js";
import { AiProviderError, type AiProvider } from "../services/ai-provider.js";
import { MockAiProvider } from "../services/mock-ai-provider.js";
import type { MealImageStorage } from "../services/meal-image-storage.js";
import { toApiMeal } from "./journal-presenter.js";

export const registerScanRoutes = async (
  app: FastifyInstance,
  repository: AppRepository,
  mealImageStorage: MealImageStorage,
  aiProvider: AiProvider = new MockAiProvider(),
): Promise<void> => {
  app.get("/v1/quota", async () => repository.getQuota());

  app.post("/v1/scans/prepare", async (_request, reply) => {
    const scan = await repository.prepareScan();
    return reply.status(201).send({
      scanId: scan.id,
      status: "prepared",
      quota: await repository.getQuota(),
    });
  });

  app.post("/v1/scans/:id/analyze", async (request, reply) => {
    const params = request.params as { id: string };
    const scan = await repository.getScan(params.id);
    if (!scan) return reply.status(404).send({ error: "scan_not_found" });
    if (scan.analyzedResponse) return scan.analyzedResponse;

    const parsed = analyzeScanRequestSchema.safeParse(request.body ?? {});
    if (!parsed.success) {
      return reply.status(400).send({
        error: "invalid_scan_image",
        issues: parsed.error.issues,
      });
    }

    const decision = decideScanQuota(await repository.getQuota());
    if (!decision.allowed) {
      return reply.status(402).send({
        error: "scan_credit_required",
        reason: decision.reason,
        quota: await repository.getQuota(),
      });
    }

    const userHint = parsed.data.hint?.trim() || undefined;
    const scanWithRequestContext = {
      ...scan,
      status: "analyzing" as const,
      userHint,
      imageMimeType: parsed.data.image?.mimeType,
      imageByteSize: parsed.data.image?.byteSize,
    };
    await repository.updateScan(scanWithRequestContext);

    let analyzedResult;
    try {
      analyzedResult = await aiProvider.analyzeMealImage({
        scanId: scan.id,
        userHint,
        image: parsed.data.image,
      });
    } catch (error) {
      await repository.updateScan({
        ...scanWithRequestContext,
        status: "failed",
      });

      if (error instanceof AiProviderError) {
        return reply.status(error.statusCode).send({
          error: error.code,
          message: error.message,
          retryable: error.retryable,
        });
      }

      return reply.status(502).send({
        error: "ai_provider_failed",
        message: "Food analysis failed.",
        retryable: true,
      });
    }

    await repository.consumeCredit(decision.reason);

    await repository.updateScan({
      ...scanWithRequestContext,
      status: "ready_for_review",
      creditReason: decision.reason,
      analyzedResponse: analyzedResult.analysis,
      aiProviderRun: analyzedResult.providerRun,
    });

    return analyzedResult.analysis;
  });

  app.post("/v1/scans/:id/confirm", async (request, reply) => {
    const params = request.params as { id: string };
    const scan = await repository.getScan(params.id);
    if (!scan) return reply.status(404).send({ error: "scan_not_found" });

    const parsed = confirmScanRequestSchema.safeParse(request.body);
    if (!parsed.success) {
      return reply.status(400).send({
        error: "invalid_confirmation",
        issues: parsed.error.issues,
      });
    }

    const image = parsed.data.image;
    const imageBytes = image ? Buffer.from(image.base64, "base64") : undefined;
    if (image && imageBytes?.byteLength !== image.byteSize) {
      return reply.status(400).send({ error: "scan_image_size_mismatch" });
    }

    let meal = await repository.createMeal({
      profileId: scan.profileId,
      mealType: parsed.data.mealType,
      title: parsed.data.title,
      source: "ai_scan",
      scanSessionId: scan.id,
      items: parsed.data.items.map((item) => ({
        displayName: item.name,
        portion: {
          quantity: item.quantity,
          unit: item.unit,
          grams: item.estimatedGrams,
        },
        nutrition: item.nutrition,
      })),
    });

    if (image && imageBytes && mealImageStorage.enabled) {
      try {
        const storedImage = await mealImageStorage.uploadMealImage({
          profileId: scan.profileId,
          mealId: meal.mealId,
          bytes: imageBytes,
          mimeType: image.mimeType,
        });
        meal = (await repository.attachMealImage(meal.mealId, storedImage)) ?? meal;
      } catch (error) {
        request.log.error(
          { err: error, mealId: meal.mealId, scanId: scan.id },
          "meal image upload failed",
        );
      }
    }

    await repository.updateScan({ ...scan, status: "confirmed" });

    return reply.status(201).send({
      mealId: meal.mealId,
      totals: meal.totals,
      meal: await toApiMeal(scan.profileId, meal, mealImageStorage),
    });
  });
};