# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LogMyPlate API
  version: 0.0.0
  description: |
    Public HTTP API for LogMyPlate, a camera-first nutrition journal.

    LogMyPlate is anonymous-first. Most user-scoped endpoints resolve identity in this order:
    1. `Authorization: Bearer <token>` for a signed-in session
    2. `x-logmyplate-install-id` for an anonymous device profile
    3. a fallback demo profile when neither is present

    The current backend reads authentication context but does not yet enforce a hard
    authenticated-only boundary on most endpoints. This document therefore marks those
    endpoints as supporting optional identity rather than requiring login.

    Every mutating endpoint except `/v1/auth/*` requires an `Idempotency-Key` header.
servers:
  - url: https://logmyplate-api.vercel.app
    description: Production
  - url: http://127.0.0.1:4000
    description: Local development
...
paths:
...
  /v1/journal/today:
    get:
      tags: [Journal]
      summary: Get today's journal
      description: Returns today's meals and nutrition totals for the current profile.
      security:
        - bearerAuth: []
        - installId: []
        - {}
      responses:
        "200":
          description: Today's journal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayJournalResponse"
  /v1/journal/range:
    get:
      tags: [Journal]
      summary: Get journal range
      description: Returns a rolling journal window plus aggregate summary metrics.
      security:
        - bearerAuth: []
        - installId: []
        - {}
      parameters:
        - name: days
          in: query
          required: false
          description: Number of days in the returned window.
          schema:
            type: integer
            minimum: 1
            maximum: 31
            default: 7
        - name: weekOffset
          in: query
          required: false
          description: Number of 7-day windows to move backward from the current week.
          schema:
            type: integer
            minimum: 0
            maximum: 104
            default: 0
      responses:
        "200":
          description: Journal range.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalRangeResponse"
        "400":
          description: Invalid range query.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /v1/journal/weeks:
    get:
      tags: [Journal]
      summary: List available journal weeks
      description: Returns week buckets that contain tracked meals, plus the current week.
      security:
        - bearerAuth: []
        - installId: []
        - {}
      responses:
        "200":
          description: Available journal weeks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalWeeksResponse"
  /v1/meals:
    post:
      tags: [Meals]
      summary: Create meal
      description: Creates a manually logged meal for the current profile.
      security:
        - bearerAuth: []
        - installId: []
        - {}
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealRequest"
      responses:
        "201":
          description: Meal created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "400":
          description: Invalid meal payload or missing idempotency key.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ValidationError"
                  - $ref: "#/components/schemas/ErrorWithMessage"
  /v1/meals/{id}:
    get:
      tags: [Meals]
      summary: Get meal
      description: Returns one meal belonging to the current profile.
      security:
        - bearerAuth: []
        - installId: []
        - {}
      parameters:
        - $ref: "#/components/parameters/MealIdPath"
      responses:
        "200":
          description: Meal found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "404":
          $ref: "#/components/responses/MealNotFound"
    patch:
      tags: [Meals]
      summary: Update meal
      description: Replaces editable meal details while preserving the original logged timestamp.
      security:
        - bearerAuth: []
        - installId: []
        - {}
      parameters:
        - $ref: "#/components/parameters/MealIdPath"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMealRequest"
      responses:
        "200":
          description: Meal updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "400":
          description: Invalid meal update or missing idempotency key.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ValidationError"
                  - $ref: "#/components/schemas/ErrorWithMessage"
        "404":
          $ref: "#/components/responses/MealNotFound"
    delete:
      tags: [Meals]
      summary: Delete meal
      description: Deletes a meal. If the meal has a stored image, the image is deleted first.
      security:
        - bearerAuth: []
        - installId: []
        - {}
      parameters:
        - $ref: "#/components/parameters/MealIdPath"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      responses:
        "204":
          description: Meal deleted.
        "400":
          description: Missing idempotency key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorWithMessage"
        "404":
          $ref: "#/components/responses/MealNotFound"
        "503":
          description: Meal image storage is not configured, so delete cannot safely complete.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorWithMessage"
        "502":
          description: Stored meal image could not be deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorWithMessage"
  /v1/quota:
    get:
      tags: [Scans]
      summary: Get scan quota
      description: Returns remaining free, rewarded, and premium scan credits for the current profile.
      security:
        - bearerAuth: []
        - installId: []
        - {}
      responses:
        "200":
          description: Current quota.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quota"
  /v1/scans/prepare:
    post:
      tags: [Scans]
      summary: Prepare scan session
      description: Creates a scan session before image analysis begins.
      security:
        - bearerAuth: []
        - installId: []
        - {}
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      responses:
        "201":
          description: Scan session prepared.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrepareScanResponse"
        "400":
          description: Missing idempotency key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorWithMessage"
  /v1/scans/{id}/analyze:
    post:
      tags: [Scans]
      summary: Analyze scan
      description: |
        Runs food-photo analysis for a prepared scan. The caller may include a short human hint
        and an image. On success, the scan consumes one available credit and returns editable
        analysis output for review.
      security:
        - bearerAuth: []
        - installId: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ScanIdPath"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeScanRequest"
      responses:
        "200":
          description: Analysis ready for review. If the scan was already analyzed, the cached analysis is returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeScanResponse"
        "400":
          description: Invalid scan image payload or missing idempotency key.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ValidationError"
                  - $ref: "#/components/schemas/ErrorWithMessage"
        "402":
          description: No available scan credit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanCreditRequiredError"
        "404":
          $ref: "#/components/responses/ScanNotFound"
        "502":
          description: AI provider failed or returned invalid output.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProviderError"
        "503":
          description: AI provider is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProviderError"
        "504":
          description: AI provider timed out.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProviderError"
  /v1/scans/{id}/confirm:
    post:
      tags: [Scans]
      summary: Confirm scan
      description: |
        Persists the reviewed AI result as a meal. If an image is supplied and image storage
        is configured, the image is uploaded and attached to the resulting meal.
      security:
        - bearerAuth: []
        - installId: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ScanIdPath"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmScanRequest"
      responses:
        "201":
          description: Scan confirmed and meal created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmScanResponse"
        "400":
          description: Invalid confirmation payload, image size mismatch, or missing idempotency key.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ValidationError"
                  - $ref: "#/components/schemas/SimpleError"
                  - $ref: "#/components/schemas/ErrorWithMessage"
        "404":
          $ref: "#/components/responses/ScanNotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque session token
      description: Optional signed-in session token.
    installId:
      type: apiKey
      in: header
      name: x-logmyplate-install-id
      description: Optional anonymous device identifier used to resolve or create a device profile.
  parameters:
...
    MealIdPath:
      name: id
      in: path
      required: true
      description: Meal ID.
      schema:
        type: string
    ScanIdPath:
      name: id
      in: path
      required: true
      description: Scan session ID.
      schema:
        type: string
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Required for mutating requests outside `/v1/auth/*`; repeated keys replay the cached response.
      schema:
        type: string
        minLength: 1
      example: 5f8b5b42-1844-4d12-bf3e-0f4208652f0b
...
  responses:
...
    MealNotFound:
      description: Meal was not found for the current profile.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SimpleError"
          examples:
            default:
              value:
                error: meal_not_found
    ScanNotFound:
      description: Scan was not found for the current profile.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SimpleError"
          examples:
            default:
              value:
                error: scan_not_found
...
  schemas:
...
    MacroTotals:
      type: object
      required: [calories, proteinG, carbsG, fatG]
      properties:
        calories:
          type: number
          minimum: 0
        proteinG:
          type: number
          minimum: 0
        carbsG:
          type: number
          minimum: 0
        fatG:
          type: number
          minimum: 0
        fiberG:
          type: number
          minimum: 0
        sugarG:
          type: number
          minimum: 0
        sodiumMg:
          type: number
          minimum: 0
    PortionUnit:
      type: string
      enum:
        - gram
        - ml
        - piece
        - serving
        - bowl
        - katori
        - cup
        - tablespoon
        - teaspoon
        - ladle
        - roti
        - idli
        - dosa
        - slice
        - scoop
        - small
        - medium
        - large
...
    MealItem:
      type: object
      required: [id, displayName, quantity, unit, grams, nutrition, userEdited]
      properties:
        id:
          type: string
        foodId:
          type: string
        displayName:
          type: string
        quantity:
          type: number
          exclusiveMinimum: 0
        unit:
          $ref: "#/components/schemas/PortionUnit"
        grams:
          type: number
          minimum: 0
        nutrition:
          $ref: "#/components/schemas/MacroTotals"
        userEdited:
          type: boolean
    MealImage:
      type: object
      required: [url, mimeType, byteSize]
      properties:
        url:
          type: string
          format: uri
        mimeType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        byteSize:
          type: integer
          minimum: 1
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
    Meal:
      type: object
      required: [id, profileId, mealType, title, loggedAt, items, totals]
      properties:
        id:
          type: string
        profileId:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, snack, dinner]
        title:
          type: string
        loggedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/MealItem"
        totals:
          $ref: "#/components/schemas/MacroTotals"
        image:
          $ref: "#/components/schemas/MealImage"
    CreateMealItem:
      type: object
      required: [displayName, quantity, unit, grams, nutrition]
      properties:
        foodId:
          type: string
        displayName:
          type: string
        quantity:
          type: number
          exclusiveMinimum: 0
        unit:
          $ref: "#/components/schemas/PortionUnit"
        grams:
          type: number
          minimum: 0
        nutrition:
          $ref: "#/components/schemas/MacroTotals"
        userEdited:
          type: boolean
          default: false
    CreateMealRequest:
      type: object
      required: [mealType, title, items]
      properties:
        mealType:
          type: string
          enum: [breakfast, lunch, snack, dinner]
        title:
          type: string
        loggedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateMealItem"
    UpdateMealRequest:
      type: object
      required: [mealType, title, items]
      properties:
        mealType:
          type: string
          enum: [breakfast, lunch, snack, dinner]
        title:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateMealItem"
    TodayJournalResponse:
      type: object
      required: [date, timezone, totals, meals]
      properties:
        date:
          type: string
        timezone:
          type: string
        totals:
          $ref: "#/components/schemas/MacroTotals"
        target:
          $ref: "#/components/schemas/MacroTotals"
        meals:
          type: array
          items:
            $ref: "#/components/schemas/Meal"
    JournalDay:
      type: object
      required: [date, mealCount, totals, meals]
      properties:
        date:
          type: string
        mealCount:
          type: integer
          minimum: 0
        totals:
          $ref: "#/components/schemas/MacroTotals"
        meals:
          type: array
          items:
            $ref: "#/components/schemas/Meal"
    JournalRangeSummary:
      type: object
      required:
        - windowDays
        - activeDays
        - mealCount
        - totals
        - trackedDayAverage
        - calendarDayAverage
      properties:
        windowDays:
          type: integer
          minimum: 1
        activeDays:
          type: integer
          minimum: 0
        mealCount:
          type: integer
          minimum: 0
        totals:
          $ref: "#/components/schemas/MacroTotals"
        trackedDayAverage:
          $ref: "#/components/schemas/MacroTotals"
        calendarDayAverage:
          $ref: "#/components/schemas/MacroTotals"
    JournalRangeResponse:
      type: object
      required: [startDate, endDate, timezone, days, summary]
      properties:
        startDate:
          type: string
        endDate:
          type: string
        timezone:
          type: string
        target:
          $ref: "#/components/schemas/MacroTotals"
        days:
          type: array
          items:
            $ref: "#/components/schemas/JournalDay"
        summary:
          $ref: "#/components/schemas/JournalRangeSummary"
    JournalWeekOption:
      type: object
      required: [weekOffset, startDate, endDate, activeDays]
      properties:
        weekOffset:
          type: integer
          minimum: 0
        startDate:
          type: string
        endDate:
          type: string
        activeDays:
          type: integer
          minimum: 0
    JournalWeeksResponse:
      type: object
      required: [weeks]
      properties:
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/JournalWeekOption"
    Quota:
      type: object
      required: [freeRemaining, rewardedRemaining, premiumRemaining]
      properties:
        freeRemaining:
          type: integer
          minimum: 0
        rewardedRemaining:
          type: integer
          minimum: 0
        premiumRemaining:
          type: integer
          minimum: 0
    PrepareScanResponse:
      type: object
      required: [scanId, status, quota]
      properties:
        scanId:
          type: string
        status:
          type: string
          enum: [prepared, analyzing, needs_ad, ready_for_review, confirmed, cancelled, failed]
        quota:
          $ref: "#/components/schemas/Quota"
    ScanImage:
      type: object
      required: [mimeType, base64, byteSize]
      properties:
        mimeType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        base64:
          type: string
          minLength: 1
          maxLength: 8000000
          description: Base64-encoded image bytes.
        byteSize:
          type: integer
          minimum: 1
          maximum: 6000000
    AnalyzeScanRequest:
      type: object
      required: [hint]
      properties:
        hint:
          type: string
          minLength: 1
          maxLength: 280
          description: Optional human clue about the plate; current schema requires a non-empty value.
        image:
          $ref: "#/components/schemas/ScanImage"
    AnalyzedMealItem:
      type: object
      required:
        - id
        - name
        - aliases
        - quantity
        - unit
        - estimatedGrams
        - preparation
        - confidence
        - nutrition
      properties:
        id:
          type: string
        name:
          type: string
        aliases:
          type: array
          items:
            type: string
        quantity:
          type: number
          exclusiveMinimum: 0
        unit:
          $ref: "#/components/schemas/PortionUnit"
        estimatedGrams:
          type: number
          minimum: 0
        preparation:
          type: string
          enum: [home, restaurant, packaged, unknown]
        confidence:
          type: number
          minimum: 0
          maximum: 1
        nutrition:
          $ref: "#/components/schemas/MacroTotals"
    AnalyzeScanResponse:
      type: object
      required: [scanId, status, mealType, mealName, detectedLanguage, items, totals]
      properties:
        scanId:
          type: string
        status:
          type: string
          const: ready_for_review
        mealType:
          type: string
          enum: [breakfast, lunch, snack, dinner]
        mealName:
          type: string
        detectedLanguage:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/AnalyzedMealItem"
        totals:
          $ref: "#/components/schemas/MacroTotals"
    ConfirmScanItem:
      type: object
      required: [name, quantity, unit, estimatedGrams, nutrition]
      properties:
        name:
          type: string
        quantity:
          type: number
          exclusiveMinimum: 0
        unit:
          $ref: "#/components/schemas/PortionUnit"
        estimatedGrams:
          type: number
          minimum: 0
        nutrition:
          $ref: "#/components/schemas/MacroTotals"
    ConfirmScanRequest:
      type: object
      required: [mealType, title, items]
      properties:
        mealType:
          type: string
          enum: [breakfast, lunch, snack, dinner]
        title:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ConfirmScanItem"
        image:
          $ref: "#/components/schemas/ScanImage"
    ConfirmScanResponse:
      type: object
      required: [mealId, totals, meal]
      properties:
        mealId:
          type: string
        totals:
          $ref: "#/components/schemas/MacroTotals"
        meal:
          $ref: "#/components/schemas/Meal"
...
    SimpleError:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ErrorWithMessage:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
    ValidationIssue:
      type: object
      additionalProperties: true
    ValidationError:
      type: object
      required: [error, issues]
      properties:
        error:
          type: string
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
    ScanCreditRequiredError:
      type: object
      required: [error, reason, quota]
      properties:
        error:
          type: string
          const: scan_credit_required
        reason:
          type: string
          enum: [needs_rewarded_ad, exhausted]
        quota:
          $ref: "#/components/schemas/Quota"
    AiProviderError:
      type: object
      required: [error, message, retryable]
      properties:
        error:
          type: string
        message:
          type: string
        retryable:
          type: boolean